rename: qr 관련 클래스들 qr 패키지 아래로 이동 - #135
